integration: remove sl tracing
Summary:
# Context

This is way too much tracing for our integration tests. It makes it very difficult to debug Eden related issues.

# This diff

Removes the tracing setting from this particular location. It might be worth having an option to set a less-verbose log level somewhere in the code, but I'll just remove it entirely for now.

Reviewed By: muirdm

Differential Revision: D59651692

fbshipit-source-id: 2222bacc0972ff63c6cc763b56bd35d91d31bfd4