General logging for expensive queries
Summary:
This diff adds a new Scuba log tracking general expensive queries

It uses the same schema as suffix_glob

Motivation
Since we are looking to extend suffix globs away from only **/* format, we would like to track what requests in general can be optimized with offloading. In order to do that we need to know what queries are expensive.

Reviewed By: jdelliot

Differential Revision: D59490175

fbshipit-source-id: 962301628ca6b5b84c91bf0d19a8d1c7432ec79d