Add offloadable but run locally duration log
Summary:
This diff adds a new ODS log "thrift.EdenServiceHandler.glob_files.local_offloadable_duration_us".
This log tracks the duration of a local run that could have been offloaded to the remote server.

Updates SuffixGlob scuba entry to track if the request was offloaded or not. Also no longer logs when there is a relative root, to reduce log noise for non-offloadable queries

Motivation

We want to be able to establish the difference between an offloaded request and a local request.
The existing log "thrift.EdenServiceHandler.glob_files.local_duration_us" tracks not only requests that could be offloaded but also every other request, so it can't really be used for this purpose.

Reviewed By: jdelliot

Differential Revision: D59349983

fbshipit-source-id: 154987a98d0c2c85eebacc15eca59326bcb25a81