Update tests to support both old and new implementations
Summary:
In the Configerator world we can set the push redirection even before the repo definition exists. In the XDB world we won't be able to.
This diff rearranges the setup of tests so that they can work both ways.

The next diff will show it working for XDB; I split them because there are code changes involved.

Reviewed By: mitrandir77

Differential Revision: D59600620

fbshipit-source-id: 720b2b1b629a6bf2016b8a0a0dd21c59ffe4a2b9