Fix rust compiler warnings in mononoke codebase
Summary:
I think they updated the compiler version and this new version has new ways to detect dead code.
It complains when a public field is never read.

Most of these were in the OSS build, so I typechecked the whole codebase in both modes.

I removed some fields that were unused, but added `#[allow(dead_code)]` to a few because:
- They look like might be used in `Debug` output
- They're from a codebase that has been created recently (e.g. Mononoke Git server).

Reviewed By: RajivTS

Differential Revision: D59671640

fbshipit-source-id: 33f74ac6e2750dbb0aab4cc7fb6aa086174fb16e