derivation_queue: streamify dequeue response
Summary:
Right now when dequeuing items out of the derivation queue the following happens:
1) We make two requests to zelos to figure out what items are ready to derive and what items are already being derived.
2) From all ready items that are not already being derived, we choose `limit` of them and we try to claim derivation for all of them concurrently.
3) After we finish claiming derivations, we iterate over items that we successfully managed to claim derivation for and we make a request for each item sequentially to read it's data from zelos.
4) After we finish reading item data, we finally spawn tasks to derive each of the items.

Waiting to claim all derivations and afterwards reading each item data sequentially in step (3) is inefficient. Instead, this diff changes the logic so that we concurrently try to claim derivation and read item data, and changes the return type of the `dequeue` method so that we start spawning derivation right after we claim derivation and read the data for an item.

Reviewed By: RajivTS

Differential Revision: D59630464

fbshipit-source-id: 80046b8a1c0aa5cc5a89e594a4f3876417b23560