Move function that gets all submodule deps to git_submodules module
Summary:
## This stack
See T188516011 for more context, but basically we currently have some duplication of the code/logic that loads the submodule dependencies for a commit sync operation.

This stack will reduce this duplication and clean up a few things.

## This diff
**Refactor/no-op**
Moves the function that will be used by all the callsites to the `git_submodules::utils` module.

Reviewed By: mitrandir77

Differential Revision: D58582519

fbshipit-source-id: 91f896d6c5a118f9505ba48918204ea2f48fd5c8