Consider submodule deps from both source and target repos
Summary:
## This stack
See D58582519

## This diff
We'll now start work to support backsyncing changes to the submodule expansion.
In order to backsync changes, we need to load submodule dependencies as well.
So instead of passing only the source repo to get all submodule dependencies, let's pass both source and target, because then all callsites don't need to worry about determining which is small and which is large.

The large repo will never have submodule dependencies anyway.

Reviewed By: mitrandir77

Differential Revision: D59670860

fbshipit-source-id: c062e8f55d2b1daadf4b186f47bdd691c80df986