use Journal to decide if use cache for getScmStatusV2
Summary:
We peek Journal to check the head sequence number. If it's a sequence number we've seen before, we can save the repeated unnecessary runs of `diff` which walks through the entire tree.

Currently as a first milestone, this behavior is hidden behind EdenConfig `hg:cache-sl-status-results`.

Added counters to track the cache hit/miss. See test plan for example output.™

Reviewed By: kmancini

Differential Revision: D59356329

fbshipit-source-id: 4d1ca4ee3ea00c965661f1731dd032a6e1efd9a9