add block_merge_commits hook
Summary:
This is to replicate https://www.internalfb.com/code/opsfiles/[master]/chef/cookbooks/other/fb_git/files/default/githooks/hooks/update.d/103-no_merges.sh from git.

### Notable differences:

git hook can be enabled and in itself it has it's own config that once again controls whether it runs or not:
https://www.internalfb.com/code/opsfiles/[16595240c2e8c763486aa4ebced7329461e0cea3]/chef/cookbooks/other/fb_git/files/default/githooks/hooks/update.d/103-no_merges.sh?lines=21-24
In Mononoke we will either configure it or not - the hook won't have its own, private config that turns it off or on.

--------

The hook has this logic that's needed for wa repos:

https://www.internalfb.com/code/opsfiles/[16595240c2e8c763486aa4ebced7329461e0cea3]/chef/cookbooks/other/fb_git/files/default/githooks/hooks/update.d/103-no_merges.sh?lines=47-50

If we didn't have to replicate it, the hook would be much simpler. Just an if condition checking if commit is a merge commit.

Reviewed By: mitrandir77

Differential Revision: D59745160

fbshipit-source-id: 66e635073b38dfa9afefb5fa7e8f7c99266b1f57