Derived data: present a unified interface for derivation that's got it all
Summary:
We've been working towards this for a long time: an efficient way to derive data that is also convenient to use:

It's similar to `derive_exactly_batch` in terms of performance, but it removes a lot of care that was previously needed to call that function:

The new API:
* Derives dependent types if necessary
* Derives ancestors of batch if necessary
* Derives multiple heads (no need to provide a batch in topological order)
* Batch size is configurable, but sane defaults can also be read from configerator if not explicitely provided
* Does only the work needed, not more: if something is already derived, and we are not asking for `rederive`, don't even try to derive it

The semantics for this function in a nutshell are:
"I need these heads derived. Please, figure out the details and be efficient about it".

We also use this abstraction to derive dependent types instead of `derive_exactly_batch`. This means that if we want to derive a type and its dependencies, we will stop walking already derived dependencies.

This is important in the context of non-roundtrippable commits and Live Git Sync as it means that one doesn't need to be careful about deriving non-git types, then writing bonsai git mappings, then deriving git types anymore. As long as we derive after writing the mappings, everything will work. See code removed in gitimport in the rest of this stack for an illustration.

Reviewed By: YousefSalama

Differential Revision: D59588693

fbshipit-source-id: e81c95abced5ae9a4ed0728e19b42f5cde77636b