Optimize Live Git Sync by creating stacks of changesets at a time
Summary:
Use `repo_create_stack` to create an entire stack and derive it in a single SCS call.

The implementation on the SCS side can use optimizations such as batched blobstore writes to provide much better performance from the prior implementation.

We had to rewrangle some of the APIs in the git import libraries to make this possible as we now don't have access to the bonsai changeset ids as early as we used to, so we need to cut any early use of it.

We also decided to reduce the logging to only show the last element of each batch to reduce log spew.

Note: We do a little less scuba logging because we lost some details on the commit at the time we acquire the ChangesetId it gets mapped to.
This information can be retrieved by inspecting the commits themselves, so I don't think it's worth jumping through hoops to restore it.

Reviewed By: RajivTS

Differential Revision: D59588688

fbshipit-source-id: 76a6544ef98f5506c95a576635c9ed958a17b711