scs: repo_create_stack: Fix subtle bug with non-roundtrippable commits
Summary:
If we're given a Git hash for a commit to create, the source of truth for this commit is its original git commit.
This means that the bonsai_git_mapping must be written to before we trigger any derivation for `GitCommits`.

This is not a bug that has manifested itself yet, but as we start relying on `repo_create_stack` for Live Git Sync in the next diff, it will become a problem if it is not addressed.

Reviewed By: mitrandir77

Differential Revision: D59588689

fbshipit-source-id: 92f22527cd21f1abdf321eb2b8db25ab956e7be0