drawdag: Use `derive_heads` instead of `derive_exactly_batch`
Summary:
`derive_heads` is designed to actually derive everything that's needed without pre-conditions on the input.

This is what we want in the context of drawdag.

This will be useful as we'll go back to tightening the contract of `derive_exactly_batch` now that `derive_heads` provides this convenient interface.

Reviewed By: YousefSalama

Differential Revision: D59751344

fbshipit-source-id: 1cc0ce8158d06a0daef4611d2cfd7bf54a694d6b