fix du on FilteredFS mounts
Summary:
# Context

`eden du` would fail on FilteredFS mounts that had filters enabled. This is because getScmStatusV2 expects the active filter to be passed in so that it can verify that Eden and Sapling agree on which commit/filter is checked out.

When the active filter isn't passed in, getScmStatusV2 fails with the following error:

```
Error: EdenService::getScmStatusV2 failed with ex(EdenError)

Caused by:
    EdenError: error computing status: requested parent commit is out-of-date: requested fae944c603acb49c4c5cb70a691acfd8bf221d5c, but current parent commit is fae944c603acb49c4c5cb70a691acfd8bf221d5c.
    Try running `eden doctor` to remediate
```

# This diff

This diff fixes `eden du` so that it passes in the active filter to getScmStatusV2 when a filter is activated.

Reviewed By: genevievehelsel

Differential Revision: D59689603

fbshipit-source-id: 42ff51026e9ee2bdb62b483914e5c52c430d933d