Use mononoke_api::Repo in GRit server instead of a custom facet container
Summary:
A key part of supporting git pushes is the ability to move bookmarks, cross repo sync commits and invoke hooks. All this functionality is currently tied to `RepoContext` which is defined specifically for facet container `mononoke_api::Repo`. Since we need this functionality in the git server at all costs, lets use the `mononoke_api::Repo` facet container instead of a custom one.

NOTE: Ideally, `RepoContext` should have been defined on a set of traits but right now the functionality is so much bundled together that it won't be an easy job to get them apart. We can do that in the future.

Differential Revision: D59751470

fbshipit-source-id: 69dfcca046ceb179c873a27c15681c44fb23cf0d