commit: stop reading .hg/branch
Summary:
Stop reading .hg/branch at commit time. This is where the current branch is stored (and then saved in commit extras). The "sl branch" command already doesn't write to it, but it could still be written by checking out a commit that has a branch set in extras.

I'n not expecting a lot of test fallout - should only be tests that have bundles saved off with branches set in the commits.

If this goes well, I'll start removing `branch()` related logic.

I started looking at removing branch because test-setdiscovery.t was failing w/ the Rust checkout because the Rust checkout does not write out .hg/branch (and it shouldn't, since we don't support hg branches).

Reviewed By: quark-zju

Differential Revision: D59588259

fbshipit-source-id: 9981c5df0254c39d8b352f0e434794b9a3dad4a3