goto: fix overwriting of symlink files on Windows
Summary:
When checkout is overwriting a symlink, it needs to be careful not to write through the symlink. For unix, we open the file using O_NOFOLLOW, but AFAIK Windows does not offer that option. Fix for Windows by removing the symlink before re-writing it.

I piped the knowledge that the file was previously a symlink through a bit (to avoid needing to lstat the file before writing).

Reviewed By: quark-zju

Differential Revision: D59588258

fbshipit-source-id: b08eee035967105a7b1a07de91f8b988ba54b1c0