Setup jest tests for components/ dir
Summary: Add jest to run tests for the components dir. This isn't super necessary, but I think it's nice to have tests set up for when we have something we want to test more thoroughly. For example, we could have isolated non-isl tests for tooltips. This also makes sure we can render all the components in a basic way without error.

Reviewed By: quark-zju

Differential Revision: D59690102

fbshipit-source-id: a5ad22512e15b1efa365f915d03107e3cb0738ac