Add ButtonGroup, use for split suggestion
Summary:
Add a container that visually groups buttons together.

This polishes a recent use of Buttons squished next to each other.

We want to remove the border radius from buttons that touch each other, and also add a slight border between them.

The API for this is quite simple, just put some Buttons in a `<ButtonGroup>`. The only annoyance is that you need to set `<ButtonGroup icon>` if your `<Button>`s  have `icon` (this is because we can't read children props from a parent, even in CSS)

Reviewed By: benmonro

Differential Revision: D59648703

fbshipit-source-id: ffc27895110c0e9da3ed261428618de1320971c8