fix Windows subprocess creation
Summary:
# Context

On Windows, process creation was consistently failing with the following error:

```
class std::system_error: CreateProcess("C:\tools\eden\libexec\edenfsctl.exe" "--etc-eden-dir" "C:\ProgramData\facebook\eden" "redirect" "fixup" "--mount" "C:\open\fbsource") failed with err code: 87
```

Where error code 87 means "The parameters are incorrect"

It was unclear why this was happening. After hours of debugging, I've found that you can fix the problem by setting the stdin, stderr, and stdout of the subprocess to null. Something must be wrong with the way we handle inherriting default values for stdin/stderr, but the root cause is not entirely clear.

While I still don't fully understand the reason CreateProcess fails, it feels like this fix is safe enough to deploy for now.

# This diff

This diff sets the stderr, stdin, and stdout handles to null on Windows. It also ensures that a window is not created for the subprocess.

Reviewed By: quark-zju

Differential Revision: D59783277

fbshipit-source-id: cfed9a190b6dc8e9a643b23f84653b3780a1d87b