Fix the XDB pushredirection implementation
Summary:
The common scenario is where both the constructor and the `get`/`set` methods are called on the small / source repo.

However some code calls `get` passing the target repo–e.g. for backsync.

In `commit_rewriting/live_commit_sync_config` I was incorrectly assuming they would be the same, and that cause the wrong entry to be loaded. Fix that.

Reviewed By: RajivTS

Differential Revision: D59642568

fbshipit-source-id: 3cfae497aa141e26c0b46d0ce545cb3a456dbd2e