Support bookmark/ref deletion during git push
Summary: The current `git push` implementation always expected a `packfile` in the request payload. However, when a push contains only ref updates pointing to null or already-known commits, then the push can be without any packfile. This diff makes the necessary change and adds an integration for ref deletion scenario

Differential Revision: D59803590

fbshipit-source-id: 91d9a695e92e6fe19d1bad7d99cbe4025aa38945