Implement basic E2E git push
Summary:
This diff marks a milestone in our implementation for `git push` in `Mononoke GRit` server. By this point, we can successfully perform a git push invoking the relevant hooks, performing the necessary permission checks and uploading the required state before eventually creating/moving/deleting the bookmark.

However, we are far from done because there are lots of missing functionalities and unhandled cases which will be supported slowly (with lots of integration tests) over the course of next few weeks.

Differential Revision: D59758039

fbshipit-source-id: 3f5941055c9b978ef1f6dcfdf2a8e19d534f24ac