Add check_fix method to FixableProblems
Summary:
# Context
Eden doctor is a tool used to try and fix eden when problems arise.

# This Diff
This diff adds a new method to FixableProblems that checks if the fix was successful. This is so we can actually determine if running the fix has done anything. Previously these would be silent failures.

# Technical Details
- Adds new method check_fix to FixableProblem
- This class is temporarily a normal class for more compact reviews; when all subclasses imlementations are done it will be converted to an abstractclass for

Differential Revision: D59655409

fbshipit-source-id: 9a45807cd62ab023316bb6508eeb5bd3f67efb59