derived_data_service: use BonsaiDerivable::Dependencies in derive_underived to avoid having to explicitly derive dependencies
Summary:
This diff makes use of the dependencies specified in the `BonsaiDerivable` trait to avoid having to explicitly call `derive_underived` for each derived data type dependency whenever a new derived data type is added.

The current code actually has a missing call to derive `SkeletonManifests` when deriving `BssmV3`. This missing call means that if the derivation of `SkeletonManifests` and `BssmV3` was initiated at close enough times, we might derive `SkeletonManifests` twice unnecessariliy. Switching to use the compile time dependencies should fix this and avoids the re-occurence of this issue in the future.

Differential Revision: D59855029

fbshipit-source-id: 94599ffe82a5e5846af3c7bcaa5d7742e9c5788f