mononoke: generate optional commit cloud ACL permission checker
Summary:
For commit cloud we need to generate a lot of dynamic ACLs, meaning we create/check the ACL depending on the edenapi request we get. This means that the ACL checker could not exist when we request it.

The implementation is the same as `aclchecker_from_identity` but simply with an `.exists()` before it

Reviewed By: RajivTS

Differential Revision: D59589869

fbshipit-source-id: 14945eebba0cacb626a5b2a61bf582d31a037314