Implement support for preventing refs pointing to blobs and trees during git push
Summary:
In Metagit, we had hook `110-no_ref_tree_blob.sh` that would be used to prevent pushing refs pointing to blobs or trees in certain repos. The reason this was a hook was because it could be disabled for repos on demand. However, in Mononoke every push corresponds to a bookmark move and an entry in `bookmarks` table. We can only add `BonsaiChangesets` in the table so if we push a ref pointing to a `tree` or a `blob`, the entire model breaks.

This diff has changes to disallow this behavior across the entire service which is actually the right thing to do

Reviewed By: mitrandir77

Differential Revision: D59863022

fbshipit-source-id: 9f5abb7981f615dad634e4c6d27b499e038fde2a