Extend the bonsai_tag_mapping API to allow for deletion
Summary: As part of `git push` responsible for deleting bookmarks, we would also need to delete tags on the remote server. In contrast to regular bookmarks (branches), tags are stored in `bookmarks` table as well as `bonsai_tag_mapping` table. We need to delete such tags from both places which is why this diff adds a method on the `bonsai_tag_mapping` facet to delete tags by name.

Reviewed By: mitrandir77

Differential Revision: D59809281

fbshipit-source-id: 926dd62aaaf4f8705366199a0c75424e1c31c1bd