Derive only git protocol specific types during git push
Summary: As part of `git push`, we import and upload git object from the input packfile but also create bonsai equivalent commits. We can optionally also derive the data for these bonsai commits before moving the bookmark. The current version of the git-server was deriving it for all configured types which is wasteful and time consuming since we only need the git protocol specific types to be derived to serve the git repo. This diff makes the necessary change

Reviewed By: mitrandir77

Differential Revision: D59810864

fbshipit-source-id: 1f4da753b857d994567759ed456e592d92cf109d