Integration test for creating/modifying tags during git push
Summary: Tags are a special snowflake in git when it comes to `git pushes`. When doing normal pushes, tags are not included by default. Updating a tag is not allowed and the only way to do so is to delete the tag on the remote server first and then push the updated tag. This diff creates an integration test to highlight these behaviors and also identifies a problem in our handling of tags which will be fixed in follow up diffs.

Reviewed By: mitrandir77

Differential Revision: D59808550

fbshipit-source-id: f7e983b4cae30d2cf25616e18688fe9fd08a114f