make run_test.sh take a build mode as argument
Summary:
I've been using this script for last few days and I find it useful so let's
expand it a little bit to support other build modes. (as suggested in comments
for D59583560)

I've chosen `--flagfile` to be consistent with name for the flag in tools like
`buck` and `arc rust-check`.

The `oss` mode has special `--oss` shortcut to encourage its usage.

Reviewed By: lmvasquezg

Differential Revision: D59858547

fbshipit-source-id: da2ff8292f5460e977b34611e78c918dc067e478