Fix upload file button
Summary:
You can click the upload file button under a text area to open the file picker.

Previously, we used `<VSCodeButton>` for this, which is implemented as a `<vscode-button>` custom element. This apparently is enough to allow the `<vscode-button>` inside a `<label>` using `htmlFor` to reference an `<input type="file">` to trigger the file picker.

However, when we switched to `<Button>`, our implementation uses an HTML `<button>`, which apparently has the limitation that it will not cascade to a `<label>`'s `for` target.

So, we need to manually trigger the `<input>`'s `onClick` via ID (which is generated by `useId` and so works fine)

Reviewed By: sggutier

Differential Revision: D59835321

fbshipit-source-id: 28e0b42e9643c28c9887c9ff6a6f07660d01cdd3