Compare the pushredirect configs we computed
Summary:
If we are using both, compare to make sure the XDB implementation and content are both correct.

The previous API used `Option` but in practice the different between `None` and `Some(false, false)` is not observable outside of this file, so it just complicates things for no gain.

Some of this is jankier than I would like (e.g. `unwrap`), but we'll be able to clean this up after we are done with the rollout i.e. next week at the latest.

Reviewed By: mitrandir77

Differential Revision: D59759705

fbshipit-source-id: d80b54b390c2ae2beca247e224857e17ab2b4221