warm_bookmarks_cache: don't return an error if a type is not enabled
Summary:
Right now warm bookmarks cache will return an error if a type is not enabled in the derived data config but is included in the list of warmers. This means that disabling a type requires a code change (in case of using hg or git warmers) and complicates adding new types. This diff changes the logic so that WBC no longer returns an error and instead makes the set of spawned warmers be the intersection of derived data config enabled types and the list of specified warmers.

overriding_review_checks_triggers_an_audit_and_retroactive_review
Oncall Short Name: scm_server_infra

Differential Revision: D59857579

Reviewed By: RajivTS

fbshipit-source-id: 7092eb4bc6874b0714c7c435c3238b4be373b168