remote_gitimport: Allow configuring justknobs in integration tests
Summary:
Prior to this change, the `remote_gitimport` executable didn't have a way to read justknobs from a file.

This meant that integration tests would ignore the provided justknobs.conf file.

With this diff, `remote_gitimport` behaves like any other executable we are testing and reads justknobs from that file.

Reviewed By: mitrandir77

Differential Revision: D59867072

fbshipit-source-id: 102b6f9c4445ad8f2ca2192da6b9c376694337b2