Live Git Sync: make the batch size configurable by jk
Summary:
Before this, the default is 20 which is probably too conservative.

From experiments with `gitimport` and `chromium/src`, concurrency 20 would allow us to process 10k commits in ~2 hours while concurrency 200 would allow us to process the same number of commits in ~1h30mins.

There is also the question of RAM and CPU usage on the metagit servers.

Long story shorts, we want the ability to quickly play around with this value without having to redeploy a new fbpkg everytime, hence the justknob.
In particular, we may need to override it for some repos if they have different characteristics, such as generally larger commits.

Reviewed By: jberenhaus

Differential Revision: D59855718

fbshipit-source-id: 516353c64f6d80776c05597585cd50e638b60d1a