Always account for loading indicator width when wrapping top bar
Summary:
The top bar of ISL shows several sets of buttons. On smaller screens, and when the commit info sidebar is open, and when the cwd is long, the left and right button sets can get close together, and sometimes wrap onto two lines.

The rightmost icon on the left side is the loading indicator, which is only sometimes visible. This leads to a jumpy UI because it will wrap the right side buttons only while the loading indicator is shown.

Instead, we should account for the space for the loading indicator, and always wrap if the loading indicator would cause wrapping when it does appear.

Reviewed By: quark-zju

Differential Revision: D59874767

fbshipit-source-id: 9b25cccafcb531f87d9f8a42efb16d5ba21eb9b3