Add slight padding between stacks in the dag
Summary:
In our dag rendering, we previously had no additional padding added between stacks, which could be considered quite "dense".

We can add an extra, small, ~6px tall set of parent lines below the link line (`|/` part). By doing this when there is only one level of ancestry, we add spacing between stacks. This makes it easier to visually parse different stacks as different.

We also adjust the size of the stack actions to ensure it is sized appropriately.

This does not affect the height of commits in a linear stack, or even in a non-linear stack.

Some notes:
- I did not test with any merge history, merely my normal commit history. I think we should add some dev tooling that simulates complicated stacks + edge cases so we can make changes like this confidently.
- we could consider making this configurable, as a "stack spacing" config, but that seems like overkill. I don't think the reduction of space is too detrimental here
- I notice this adds padding even in the case where "You are here" is above a public parent. This is perhaps undesirable, as it distances "you are here" from the public parent, which is not really true. Though, if you have uncommitted changes, then it suddenly makes sense again, as you will be branching off. So I think it's OK to leave the small padding in all these cases.

Reviewed By: sggutier

Differential Revision: D59874590

fbshipit-source-id: b5b9ad5891561a41263197c5a78c26cc83a1e86d