log transport errors explicitly
Summary:
sl occasionally sees "ThriftError: Client was dropped" errors when it tries to
talk to EdenFS.
https://fb.workplace.com/groups/thriftusers/permalink/1414738659235542/

We don't really know what they mean. Rust folks have found the root cause place
where the full error message is being dropped, and are working on a fix
(to tokio).

In the meantime we can also get more info out of hg by logging the error. Using
`error!()` doesn't seem to be captured anywhere (dev_command_timers, or
hg_errors). Let's explicitly log this column to scuba.

My plan is to set the retention for this column really really low so that it
does not effect the overall retention of the table. Not sure yet exactly what
really really low should be, so suggestions welcome.

Reviewed By: sggutier

Differential Revision: D59884418

fbshipit-source-id: 97550a2543f76d7935093df741758c6e87ee7bdc