add support for eagerepo in Eden integration tests
Summary:
# Context

The new AugmentedManifest format used by Sapling and Mononoke will support looking up Trees/Blobs from CAS via (Blake3,size) pairs. We already provide a way for users to get (Blake3,size) pairs for blobs, but there is currently no easy way to get this information for trees.

We want to add support to getAttributesFromFilesV2 for looking up the (Blake3,size) pairs for trees as well. To do this, we'll need to add a method for querying TreeMetadata (including size and Blake3 hashes) via the SaplingBackingStore.

This initial implementation of TreeMetadata will only include the following features:

1) Ability to query TreeMetadata from the SaplingBackingStore via a getTreeMetadata endpoint
2) The ability to request TreeMetadata from the getAttributesFromFilesV2 thrift endpoint

In the future, we will extend the implementation to support:

1) Automatically fetching TreeMetadata during BackingStore::getTree requests
2) Caching TreeMetadata in Eden's in-memory caches

# This diff

This diff adds the ability to use eagerepo in integration tests. It essentially gives tests the opportunity to "push" commits to eagerepo so that "remote" fetches of that pushed commit data can later be performed by eagerepo.

This is needed because the only way we can test tree aux data is by having eagerepo derive the tree aux data on demand and have the treestore "fetch" this tree aux data from eagerepo.

In other words, our tests previously assumed that all data was stored locally and that no "network" fetches could be performed. We now allow network fetches via eagerepo.

Reviewed By: muirdm

Differential Revision: D59642618

fbshipit-source-id: c942d6a88b7d5973b895b607480447663b222b6f