remote_gitimport: Add retries for all SCS calls
Summary:
SCS calls can fail. In particular, some of these are long running and the connection can drop for any number of reason, such as server restarts on the SCS side.

Let's give it 3 shots before giving up, which was already the intent based on the pre-existing `UPLOAD_RETRIES` const.

Reviewed By: lmvasquezg

Differential Revision: D59969546

fbshipit-source-id: 7111ba0f55b641796625d2ddc6f9ecf353235208