Honor pushvars bypass during non-fast-forward pushes in Mononoke Git server
Summary:
The `move_bookmark.rs` file exposes bookmark movement APIs that take a separate flag called `allow_non_fast_forward` that seemingly controls is a bookmark can undergo non-fast-forward move or not. However, that's not quite true.

This flag is meant to represent the client's opinion of whether non-ffwd moves should be allowed for a particular bookmark. If the client wants only fast forward moves, then we don't check on the server we straight away block non-ffwd moves. In git, I have the flag to `true`. Git clients by default do NOT allow non-fast-forward moves and that results in such requests never even reaching the server.

Thus, if the request for a push is already at the server then that means the client wants us to move the bookmark in non-ffwd fashion, subject to server configs. The server configuration for restricting such bookmark moves takes the shape of [`BookmarkConfig`](https://fburl.com/code/u3r6roar) with `only_fast_forward` flag controlling the outcome.

Both the client side and server side restrictions can be bypassed by using the pushvar. This diff adds support to make that happen along with interation test displaying how it can be done.

Reviewed By: mitrandir77

Differential Revision: D59963424

fbshipit-source-id: 33841cf34087bf5a27dfdebe127dd552287ced56