Use WBC without any derivation in Mononoke Git Server
Summary: The only reason for using Warm Bookmarks Cache in Mononoke Git server is to avoid the unnecessary extra load on the `bookmarks` MySQL table. Before any git commit lands in Mononoke, it either comes through Mononoke Git server or through remote Git import. In either case, we already derive GDM (and GDMv2) for them so we don't really need WBC to do that for us.

Reviewed By: YousefSalama

Differential Revision: D59963986

fbshipit-source-id: f250f10c26563cbeb3d07d0ef7057623b5488d20