repo_derived_data: allow accessing derived data managers for different configs
Summary: Right now the `RepoDerivedData` facet only supports doing operations using the enabled config name, which is fine because this facet is intended to be the main way for services to query derived data. However, I want to get rid of `DerivedDataUtils` which requires providing another way to access derived data managers for different configs. This diff changes `RepoDerivedData` to make it store derived data managers for all configs, and adds a `manager_for_config` method while keeping the current interface the same.

Differential Revision: D59914310

fbshipit-source-id: 9239c85709ad529ebd3da782db11fcce66ee85af