SCS: in `repo_create_stack`, propagate the batch size to data derivation
Summary:
When someone calls `repo_create_stack`, they configure the stack size that they pass in to get the expected performance characteristics.

Before this change, the batch size that is passed in is not propagated to data derivation, which means we don't get the
full potential from configuring the batch size.

This behaviour has been observed for a large push on chromium/src where `repo_create_stack` is called with batches of 200 commits, but we only batch derivation by up to 20 commits.

Reviewed By: YousefSalama, RajivTS

Differential Revision: D60043098

fbshipit-source-id: cae4529bf0293de58f7a4b5a0fad6468f50316f9