tree-pattern-match: handle overlapping searches
Summary:
Add a way to "cut off" search states so we can better handle overlapping
matches more correctly.

The test changes because previously when querying "what is the start of match
ending at 'z'"? It reports overlapping start and is discarded:

    a b c xx d e f xx x y z
    ^^^^^^^^^^^^
             ^^^^^^^^^^^^^^

This diff adds a "cut off" logic so the match ending at "z" can stop at "e" and
won't be discarded:

    a b c xx d e f xx x y z
    ^^^^^^^^^^^^
                 ^^^^^^^^^^

Reviewed By: zzl0

Differential Revision: D59875173

fbshipit-source-id: 5fcce7d4fbeb11a3301b21fe4bc1bbb80f709228