Make multiple split title edits use a single mutation for history
Summary:
Previously, as you type a title in the split editor, each keystroke made a new history operation, so you'd undo each keystroke one-by-one.

This is a bit silly, it makes a lot more sense to undo everything you typed in one go. It's also not efficient—we use a lot of unnecessary memory to represent all these operations for each keystroke.

We need to slightly update our model to suppor this well. We add a `replaceTopOperation()` to complement `push()`, which acts like `pop` then `push`. When we detect the last operation according to our undo history is also a metaedit for the same commit, we can replace instead of push.

Reviewed By: quark-zju

Differential Revision: D59932932

fbshipit-source-id: 45ce9d5ad0c5989545a111cad09eaf861515d716