Allow editing empty split commit titles
Summary:
Previously, in the split UI, we disabled editing split commit titles for "blank" commits with no files yet moved there.

However, some workflows would like to create blank commits, give them names, then start moving lines of code there to simplify how they think about the commits.

So this diff removes the restriction that you cannot edit commit titles for blank commits.

The original reason this was in place is because the data model did not directly support it. The blank commit shown to the right of the last commit is in a sense "fake", because it's added to the "dense substack" view of the original stack. Thus, it has no associated real commit which could be edited.

There are a few ways we could remedy this, such as saving the entered title text in a new piece of state or something. This is not ideal because then we need to do a lot of synchronization as you edit the stack to reset this title.

Instead, we'd like to lean on the data model to store this just like we do all the other commit titles.

One simple approach is to make a new "real" blank commit into the stack when you type anything into the fake blank commit. This real blank commit will have the new title, supports undo/redo, and appears in the list of commits.

Once you type into the box, it will mean that another fake blank commit is created after it. I think that's ok.

Reviewed By: quark-zju

Differential Revision: D59942286

fbshipit-source-id: 3cbbfb10507a7f06bbbd1f060d4df1d07826b1ad