Allow rolling out Eden configs using QE
Summary:
The goal of this stack of diffs is to allow rolling out EdenFS features using a QE. This will enable measuring the impact of the change in deltoid, where a number of eden, as well as downstream metrics (Meerkat, SandboxCache, etc) are defined.

In this diff, we add a call to the [QE thrift API](https://www.internalfb.com/intern/wiki/Qe2/Setup/Experimenting_in_C++/Experimenting_in_fbcode/) and thread through the values to eden configs.

There is one main drawback of the approach so far:
QEs allow you to define boolean, string, int params, and multiple values per rollout group, but we're not using those. We only support a boolean param defined in the QE, and then separately define whatever we want to pass down via `ConfigItem` (and use the QE only for gating). This is certainly doable, but needs a fair bit of refactor; and the current diff seems like a good MVP to test out whether QEs are actually helpful

Reviewed By: genevievehelsel

Differential Revision: D59507578

fbshipit-source-id: 1587e8a77221d6e39cfeabef3c34d87a58c911d9