Consider submodule deps from both source and target repos (re-land after fix)
Summary:
This is a re-land of D59670860 after D59861126 fixed the OOM issues.

D59670860 was committed after the culprit diff D58582518, so I had to revert it as well.

I'll land this on Tuesday, when I'm back from PTO and D59861126 will have been in prod for a while, so we'll be confident that the issue was fixed.

Reviewed By: RajivTS

Differential Revision: D59910169

fbshipit-source-id: 49127189d493b98c2faac8a0315c31e66f9de20b