Live Git Sync: de-pessimization: Don't require `ChangesetInfo` derivation
Summary:
This fixes a situation we noticed during test pushes with `chromium/src` in prod:
A push of 1000 commits triggered 1000 derivations of `ChangesetInfo`.
That is very bad for performance. We are going out of our way to batch derivation to avoid
these kinds of behaviours, so this defeats the purpose of previous optimizations.

It turns out, we accidentally implicitely derived `ChangesetInfo` when accessing `hg_extra`.
Instead, pass them down from the SCS request parameters.

Reviewed By: YousefSalama

Differential Revision: D60108989

fbshipit-source-id: cd14e014859043e0fd51f1a1c91324cee842ca87